MapReduce-KNN
K nearest neighbour implementation for Hadoop MapReduce

This is a java program designed to work with the MapReduce framework.
In this example the K nearest neighbour classification method
(supervised machine learning) is applied to some sample data about car
types and buyer characteristics, so that it classifies a buyer with a
likely car model.